back-end/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
screen==1.0.1
sqlparse==0.4.2
xlwt==1.3.0
